# perf(core): use BTreeMap for ResourceTable